+++ lib/src/change_emitter_base.dart
@@ -83,7 +83,7 @@ abstract class ChangeWithAny extends Change {
   ChangeWithAny({required bool quiet, required this.any}) : super(quiet: quiet);
 }
 
-abstract class ParentEmitter<C extends Change> extends ChangeEmitter<C> {
+mixin ParentEmitter<C extends Change> on ChangeEmitter<C> {
   @protected
   void registerChildren();
   @protected
@@ -125,7 +125,7 @@ abstract class ParentEmitter<C extends Change> extends ChangeEmitter<C> {
 ///
 ///
 abstract class EmitterContainer<C extends ContainerChange>
-    extends ParentEmitter<C> {
+    extends ChangeEmitter<C> with ParentEmitter {
   EmitterContainer({this.emitDetailedChanges = false});
 
   void registerChildren() {
@@ -148,20 +148,19 @@ abstract class EmitterContainer<C extends ContainerChange>
   ///override this method in order to create and use your own subclass of [ContainerChange]
   ///If you use [EmitterContainer.emit] then this function will be called with child and childChange as null
   @protected
-  C containerChangeFromChildChange(
-      {ChangeEmitter? child, Change? change, bool quiet = false}) {
-    if (emitDetailedChanges && child != null)
-      return ContainerChange(child, change) as C;
+  C containerChangeFromDependency(
+      {ChangeEmitter? dependency, Change? change, bool quiet = false}) {
+    if (emitDetailedChanges && dependency != null)
+      return ContainerChange(dependency, change) as C;
     return ContainerChange.any() as C;
   }
 
   Stream<C> _getStream() {
-    var elements = emittingChildren ?? children;
-    var streams = <Stream<C>>[];
-    streams
-      ..addAll(elements.map((e) => e.changes.where((event) => !event.quiet).map(
-          (event) => containerChangeFromChildChange(child: e, change: event))))
-      ..add(_controller.stream);
+    var streams = dependencies
+        .map((e) => e.changes.where((event) => !event.quiet).map((event) =>
+            containerChangeFromDependency(dependency: e, change: event)))
+        .toList()
+          ..add(_controller.stream);
 
     return StreamGroup.merge<C>(streams).asBroadcastStream();
   }
@@ -171,17 +170,17 @@ abstract class EmitterContainer<C extends ContainerChange>
   ///whenever any of the children change so that UI or other elements of your state
   ///can update reactively. If you only want a subset of
   ///the children to trigger changes, override [emittingChildren].
-  List<ChangeEmitter> get children;
+  Set<ChangeEmitter> get children => {};
 
   ///A list of [children] that should trigger this container to emit changes. If you want all children
   ///to trigger changes, then you don't need to override this getter.
-  List<ChangeEmitter>? get emittingChildren => null;
+  Set<ChangeEmitter> get dependencies => children;
 
   ///Emits [new ContainerChange.any]).
   ///
   ///To emit a change but prevent a parent [EmitterContainer] from emitting a change, set quiet to true.
   void emit({bool quiet = false}) =>
-      addChangeToStream(containerChangeFromChildChange(quiet: quiet));
+      addChangeToStream(containerChangeFromDependency(quiet: quiet));
 
   ///Disposes resources of all [children] and [this].
   @mustCallSuper
@@ -224,5 +223,20 @@ class ContainerChange extends ChangeWithAny {
   ///the same cached object to minimize garbage collection.
   factory ContainerChange.any({bool quiet = false}) {
     return quiet ? ContainerChange._any(quiet: true) : _anyCache;
+  }
+}
+
+abstract class ListenableEmitterMixin extends ChangeEmitter
+    implements Listenable {
+  final _listeners = <VoidCallback, StreamSubscription>{};
+
+  @override
+  void addListener(VoidCallback listener) {
+    _listeners[listener] = changes.listen((event) => listener());
+  }
+
+  @override
+  void removeListener(VoidCallback listener) {
+    _listeners.remove(listener)?.cancel();
   }
 }